+++ history.md
@@ -1,3 +1,8 @@
+[2021/12/21]
+- Add
+  + open.commons.csv.CsvHeader
+    + CsvHeader(@JsonProperty("name") String, @JsonProperty("dataType") ColumnDataType): 명시적으로 JsonProperty 추가
+
 [2021/12/15]
 - Modify
   + open.commons.csv.AbstractCsvFileLoad: REST API RequestEntity로 사용하기 위한 #headers 'Setter' 메소드 설정 추가